stopped removed redundant menu initialization

+++ app/src/main/java/com/shane/baking/adapters/RecipeAdapter.java
@@ -70,7 +70,7 @@ public class RecipeAdapter extends RecyclerView.Adapter<RecipeAdapter.RecipeView
         void bind(Recipe recipe) {
             nameTextView.setText(recipe.getName());
             servingsTextView.setText(String.valueOf(recipe.getServings()));
-            itemView.setId(recipe.getId());
+            itemView.setId((int) recipe.getId());
         }
 
         @Override
